+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "simple-html"
-version = "0.7.0"
+version = "1.0.0"
 readme = "README.md"
 description = "Template-less html rendering in Python"
 authors = ["Keith Philpott <fakekeith@example.org>"]
@@ -19,9 +19,9 @@ mypy = "1.6.0"
 pytest = "7.4.2"
 setuptools = "68.2.2"
 black = "23.9.1"
-django = "^4.2.6"
-fast-html = "^1.0.3"
-dominate = "^2.8.0"
+django = "4.2.6"
+fast-html = "1.0.3"
+dominate = "2.8.0"
 
 
 [build-system]
